renamed module to match PEP-8 suggestions
