+++ doc/examples/gerchberg_saxton.py
@@ -7,12 +7,12 @@ Created on Thu Aug  4 15:44:04 2022
 
 import numpy as np
 
-from light_prop.algorithms import GerchbergSaxton
-from light_prop.calculations import gaussian
-from light_prop.calculations import get_gaussian_distribution
-from light_prop.lightfield import LightField
-from light_prop.propagation.params import PropagationParams
-from light_prop.visualisation import GeneratePropagationPlot, PlotTypes
+from lightprop.algorithms import GerchbergSaxton
+from lightprop.calculations import gaussian
+from lightprop.calculations import get_gaussian_distribution
+from lightprop.lightfield import LightField
+from lightprop.propagation.params import PropagationParams
+from lightprop.visualisation import GeneratePropagationPlot, PlotTypes
 
 if __name__ == "__main__":
     params = PropagationParams.get_example_propagation_data()
